--- src/context/PokemonContext.js
@@ -1,4 +1,0 @@
-import { createContext } from "react";
-
-// null은 초기값
-export const PokemonContext = createContext(null);
