+++ src/pages/Dex.jsx
@@ -1,23 +1,13 @@
 import styled from "styled-components";
 import Dashboard from "../components/Dashboard";
 import PokemonList from "../components/PokemonList";
-import { useState } from "react";
-import { PokemonContext } from "../context/PokemonContext";
+import { PokemonProvider } from "../context/PokemonContext";
 
 const Dex = () => {
   // myPokemon: 사용자가 선택한 포켓몬 목록을 저장하는 배열, setMyPokemon: myPokemon 상태를 업데이트하는 함수
-  const [myPokemon, setMyPokemon] = useState([]);
-  const [pokemonList, setPokemonList] = useState(null);
 
   return (
-    <PokemonContext.Provider
-      value={{
-        myPokemon,
-        setMyPokemon,
-        pokemonList,
-        setPokemonList,
-      }}
-    >
+    <PokemonProvider>
       <RootDiv>
         <div className="app-container">
           {/* Dashboard 컴포넌트 */}
@@ -27,7 +17,7 @@ const Dex = () => {
           <PokemonList />
         </div>
       </RootDiv>
-    </PokemonContext.Provider>
+    </PokemonProvider>
   );
 };
